+++ ftplugin/c.lua
@@ -1,5 +1,5 @@
-require("lsp.clang").format()
-require("lsp.clang").lint()
-require("lsp.clang").lsp()
-require("lsp.clang").dap()
+require("lang.clang").format()
+require("lang.clang").lint()
+require("lang.clang").lsp()
+require("lang.clang").dap()
